finalize efficient hint computation for bls
